requirements.txt
python.dotenv==1.0.0
streamlit
pdfplumber
openai
